Initial Commit
Current Status:
The core lexer and sugar lexer are both implemented and working
flawlessly. adding new tokens is very easy if we need to extend the
language later

The core parser is done, and the next step is writing the sugar parser,
which shouldn't be too bad.

Have not started working on the evaluator yet.